+++ std-bits/table/src/main/java/org/enso/table/parsing/DecimalParser.java
@@ -4,9 +4,9 @@ import java.text.DecimalFormat;
 import java.text.ParsePosition;
 import org.enso.table.data.column.builder.object.Builder;
 import org.enso.table.data.column.builder.object.NumericBuilder;
-import org.enso.table.parsing.problems.NumericProblemAggregator;
+import org.enso.table.parsing.problems.ProblemAggregator;
 
-public class DecimalParser extends DatatypeParser<NumericProblemAggregator> {
+public class DecimalParser extends IncrementalDatatypeParser {
   private final String thousandsSeparator;
   private final char decimalPoint;
   private final DecimalFormat decimalFormat;
@@ -38,7 +38,7 @@ public class DecimalParser extends DatatypeParser<NumericProblemAggregator> {
   }
 
   @Override
-  public Object parseSingleValue(String text, NumericProblemAggregator problemAggregator) {
+  protected Object parseSingleValue(String text, ProblemAggregator problemAggregator) {
     if (thousandsSeparator != null
         && (text.startsWith(thousandsSeparator) || text.endsWith(thousandsSeparator))) {
       problemAggregator.reportInvalidFormat(text);
@@ -84,12 +84,7 @@ public class DecimalParser extends DatatypeParser<NumericProblemAggregator> {
   }
 
   @Override
-  public Builder makeBuilderWithCapacity(long capacity) {
+  protected Builder makeBuilderWithCapacity(long capacity) {
     return NumericBuilder.createDoubleBuilder((int) capacity);
-  }
-
-  @Override
-  public NumericProblemAggregator makeProblemAggregator() {
-    return new NumericProblemAggregator();
   }
 }