+++ CHANGELOG.md
@@ -124,6 +124,7 @@
   specified type.][3455]
 - [Promote with, take, finalize to be methods of Managed_Resource
   instance][3460]
+- [Implemented automatic type detection for `Table.parse_values`.][3462]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -192,6 +193,7 @@
 [3457]: https://github.com/enso-org/enso/pull/3457
 [3455]: https://github.com/enso-org/enso/pull/3455
 [3460]: https://github.com/enso-org/enso/pull/3460
+[3462]: https://github.com/enso-org/enso/pull/3462
 
 #### Enso Compiler
 
@@ -210,6 +212,8 @@
 - [Move Builtin Types and Methods definitions to stdlib][3363]
 - [Reduce boilerplate by generating BuiltinMethod nodes from simple method
   signatures][3444]
+- [Generate boilerplate classes related to error handling and varargs in
+  builtins from method signatures][3454]
 - [Avoid needless concatenations of warning/error messages][3465]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
@@ -230,6 +234,7 @@
 [3363]: https://github.com/enso-org/enso/pull/3363
 [3444]: https://github.com/enso-org/enso/pull/3444
 [3453]: https://github.com/enso-org/enso/pull/3453
+[3454]: https://github.com/enso-org/enso/pull/3454
 [3465]: https://github.com/enso-org/enso/pull/3465
 
 # Enso 2.0.0-alpha.18 (2021-10-12)